requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
build==1.2.2.post1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
Django==5.1.3
django-browser-reload==1.17.0
django-tailwind==3.8.0
djangorestframework==3.15.2
environs==11.2.1
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
packaging==24.2
pip-tools==7.4.1
psycopg2==2.9.10
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.3
wheel==0.45.1